requirements.txt
matplotlib
pdf2image
numpy
language-check
google-cloud-vision
re
